pyproject.toml
[project]
name = "dora-eda"
version = "3.0.1"
description = "Exploratory data analysis and presentation tool"
authors = [
    {name = "Asif Sayyed",email = "asifdotexe@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "typer[rich] (>=0.19.2,<0.20.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

urls = {Repository = "https://github.com/Asifdotexe/DORA"}

[tool.poetry]
package-mode = true
packages = [{ include = "dora", from = "src" }]

[tool.poetry.scripts]
dora = "dora.main:main"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.9"
black = "^25.9.0"
isort = "^6.1.0"
pre-commit = "^4.3.0"
ipykernel = "^6.30.1"
pytest = "^8.4.2"
pytest-benchmark = "^5.1.0"
snakeviz = "^2.2.2"

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = ["redefined-outer-name", "invalid-name", "assignment-from-no-return", "too-many-branches",
           "unused-argument", "unnecessary-pass", "fixme", "too-few-public-methods"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
